Add files via upload
This WebApp helps you to visualize and provides insights into all the Terrorist attacks, assassination, and other types of attacks from 1970 till 2018. The data can be filtered and displayed in different categories and 2 visualization tools are provided, Map tool and area Chart tool. The data is also separately visualized for India as well.